Adding Parameters to book (#12)
So....
I added Arg Parser, add a few arguments, parametrized the booking,
changed the try except so it would add notifications... blah blah.

To be honest, I didn't test it in all scenarios possible. But it does
book, when it has courts of the type specified, when it doesn't it tries
to book the other type. It allows you to set date and time, etc.

Feel free to comment so I can take another two months to modify it. I
had it ready, just saw that the feature didn't have business value
anymore, so canceled my pull request.

---------

Co-authored-by: Pierluigi Dell'Arciprete <pierluigi.dellarciprete@glovoapp.com>