Adding tz to datetime now to prevent booking wrong day to to timezone… (#15)
There was an issue where the server was in a timezone and the booking in
another, hence when getting the date and adding 10 days on top, we ended
up booking the wrong day (monday)

By setting up the timezone in the datetime.now function, we prevent this
issue from happening.

I have also removed the defaults, just in case the parser is not
correctly set up in the right time zone, but they are still covered in
the booking function.

Finally added a test that shows (if you remove the zoneinfo fix) the
error.